Merge Feat/database date mappings into entity refactor branch

See merge request cse1105/2019-2020/repos/oopp-group-30/repository-template!72

+++ server/src/main/java/nl/tudelft/oopp/server/repositories/FoodcourtRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
  * interface to implement crud operations for food court.
  */
 @Repository
-public interface FoodcourtRepository extends JpaRepository<Foodcourt, Integer> {
+public interface FoodcourtRepository extends JpaRepository<Foodcourt, Long> {
 }
